Refactor code to reduce use of templates
