+++ src/libnchg/CMakeLists.txt
@@ -66,11 +66,13 @@ target_sources(
     "${CMAKE_CURRENT_SOURCE_DIR}/expected_matrix_impl.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/expected_values_aggregator.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/expected_values_aggregator_impl.hpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/expected_values.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/expected_values_impl.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/fdr_impl.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/k_merger_impl.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/mad_max_filter.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/mad_max_filter_impl.hpp"
+    "${CMAKE_CURRENT_SOURCE_DIR}/nchg.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/nchg_impl.hpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/observed_matrix.cpp"
     "${CMAKE_CURRENT_SOURCE_DIR}/observed_matrix_impl.hpp"
