+++ src/libnchg/include/nchg/nchg.hpp
@@ -28,9 +28,9 @@ NCHG_DISABLE_WARNING_POP
 // clang-format on
 
 #include <cstdint>
-#include <hictk/cooler/cooler.hpp>
 #include <hictk/file.hpp>
-#include <hictk/hic.hpp>
+#include <hictk/genomic_interval.hpp>
+#include <hictk/pixel.hpp>
 #include <hictk/transformers/join_genomic_coords.hpp>
 #include <memory>
 
@@ -55,42 +55,41 @@ struct NCHGResult {
   bool operator!=(const NCHGResult& other) const noexcept;
 };
 
-template <typename File>
-  requires HictkSingleResFile<File>
 class NCHG {
  public:
+  using N = std::uint32_t;
+  template <typename PixelIt>
   class iterator;
   using Stats = NCHGResult;
 
  private:
-  using N = std::uint32_t;
-  using ThinPixelIt = decltype(std::declval<File>().fetch("chr1", "chr2").template begin<N>());
-  using PixelIt =
-      decltype(std::declval<hictk::transformers::JoinGenomicCoords<ThinPixelIt>>().begin());
-
-  std::shared_ptr<const File> _fp;
+  std::shared_ptr<const hictk::File> _fp;
 
   hictk::Chromosome _chrom1{};
   hictk::Chromosome _chrom2{};
 
   std::shared_ptr<const ExpectedMatrix> _exp_matrix{};
   std::shared_ptr<const ObservedMatrix> _obs_matrix{};
-  ExpectedValues<File> _expected_values{};
+  ExpectedValues _expected_values{};
 
   mutable std::vector<double> _nchg_pval_buffer{};
 
  public:
-  using Params = typename ExpectedValues<File>::Params;
-  static constexpr auto& DefaultParams = ExpectedValues<File>::DefaultParams;
-  explicit NCHG(std::shared_ptr<const File> f, const hictk::Chromosome& chrom1,
+  using IteratorVariant =
+      std::variant<iterator<hictk::cooler::PixelSelector>, iterator<hictk::hic::PixelSelector>,
+                   iterator<hictk::hic::PixelSelectorAll>>;
+
+  using Params = typename ExpectedValues::Params;
+  static constexpr auto& DefaultParams = ExpectedValues::DefaultParams;
+  explicit NCHG(std::shared_ptr<const hictk::File> f, const hictk::Chromosome& chrom1,
                 const hictk::Chromosome& chrom2, const Params& params);
-  NCHG(std::shared_ptr<const File> f, const hictk::Chromosome& chrom1,
-       const hictk::Chromosome& chrom2, ExpectedValues<File> expected_values) noexcept;
+  NCHG(std::shared_ptr<const hictk::File> f, const hictk::Chromosome& chrom1,
+       const hictk::Chromosome& chrom2, ExpectedValues expected_values) noexcept;
 
   [[nodiscard]] auto params() const noexcept -> Params;
 
-  [[nodiscard]] auto observed_matrix() const noexcept -> const ObservedMatrix&;
-  [[nodiscard]] auto expected_matrix() const noexcept -> const ExpectedMatrix&;
+  [[nodiscard]] const ObservedMatrix& observed_matrix() const noexcept;
+  [[nodiscard]] const ExpectedMatrix& expected_matrix() const noexcept;
 
   [[nodiscard]] auto compute(const hictk::GenomicInterval& range, double bad_bin_fraction) const
       -> Stats;
@@ -99,29 +98,26 @@ class NCHG {
       -> Stats;
 
   [[nodiscard]] auto cbegin(const hictk::Chromosome& chrom1, const hictk::Chromosome& chrom2) const
-      -> iterator;
+      -> IteratorVariant;
   [[nodiscard]] auto cend(const hictk::Chromosome& chrom1, const hictk::Chromosome& chrom2) const
-      -> iterator;
+      -> IteratorVariant;
 
   [[nodiscard]] auto begin(const hictk::Chromosome& chrom1, const hictk::Chromosome& chrom2) const
-      -> iterator;
+      -> IteratorVariant;
   [[nodiscard]] auto end(const hictk::Chromosome& chrom1, const hictk::Chromosome& chrom2) const
-      -> iterator;
+      -> IteratorVariant;
 
-  [[nodiscard]] auto compute_expected_profile() const
-      -> std::pair<std::vector<double>, phmap::btree_map<hictk::Chromosome, double>>;
+  [[nodiscard]] std::pair<std::vector<double>, phmap::btree_map<hictk::Chromosome, double>>
+  compute_expected_profile() const;
 
  private:
-  [[nodiscard]] static auto init_exp_matrix(const hictk::Chromosome& chrom1,
-                                            const hictk::Chromosome& chrom2, const File& fp,
-                                            const ExpectedValues<File>& expected_values)
-      -> std::shared_ptr<const ExpectedMatrix>;
-  [[nodiscard]] static auto init_obs_matrix(const hictk::Chromosome& chrom1,
-                                            const hictk::Chromosome& chrom2, const File& fp,
-                                            const std::vector<bool>& bin1_mask,
-                                            const std::vector<bool>& bin2_mask, double mad_max_,
-                                            std::uint64_t min_delta_, std::uint64_t max_delta_)
-      -> std::shared_ptr<const ObservedMatrix>;
+  [[nodiscard]] static std::shared_ptr<const ExpectedMatrix> init_exp_matrix(
+      const hictk::Chromosome& chrom1, const hictk::Chromosome& chrom2, const hictk::File& f,
+      const ExpectedValues& expected_values);
+  [[nodiscard]] static std::shared_ptr<const ObservedMatrix> init_obs_matrix(
+      const hictk::Chromosome& chrom1, const hictk::Chromosome& chrom2, const hictk::File& fp,
+      const std::vector<bool>& bin1_mask, const std::vector<bool>& bin2_mask, double mad_max_,
+      std::uint64_t min_delta_, std::uint64_t max_delta_);
 
   [[nodiscard]] static double compute_cumulative_nchg(std::vector<double>& buffer,
                                                       std::uint64_t obs, std::uint64_t N1,
@@ -141,7 +137,10 @@ class NCHG {
                                                   double min_omega = 0.1);
 
  public:
+  template <typename PixelSelector>
   class iterator {
+    using PixelIt = decltype(std::declval<PixelSelector>().template begin<N>());
+    std::shared_ptr<const PixelSelector> _sel{};
     PixelIt _pixel_it{};
     PixelIt _sentinel_it{};
 
@@ -168,11 +167,11 @@ class NCHG {
     using iterator_category = std::forward_iterator_tag;
 
     iterator() = default;
-    iterator(PixelIt pixel_it, PixelIt sentinel_it, std::shared_ptr<const ObservedMatrix> obs,
+    iterator(PixelSelector sel, std::shared_ptr<const ObservedMatrix> obs,
              std::shared_ptr<const ExpectedMatrix> exp,
              std::shared_ptr<const std::vector<bool>> bin_mask1,
              std::shared_ptr<const std::vector<bool>> bin_mask2, std::uint64_t min_delta,
-             std::uint64_t max_delta) noexcept;
+             std::uint64_t max_delta);
 
     iterator(const iterator& other);
     iterator(iterator&& other) noexcept = default;
