+++ src/nchg/expected/expected.cpp
@@ -69,7 +69,7 @@ static void process_cis_chromosomes(FilePtr f, const ExpectedConfig &c) {
   const auto mask = parse_bin_mask(f->chromosomes(), f->resolution(), c.path_to_bin_mask);
 
   using File = std::remove_cvref_t<decltype(*f)>;
-  const auto evs = ExpectedValues<File>::cis_only(
+  const auto evs = ExpectedValues::cis_only(
       f,
       {c.mad_max, c.min_delta, c.max_delta, c.bin_aggregation_possible_distances_cutoff,
        c.bin_aggregation_observed_distances_cutoff, c.interpolate_expected_values,
@@ -87,7 +87,7 @@ static void process_trans_chromosomes(FilePtr f, const ExpectedConfig &c) {
   const auto mask = parse_bin_mask(f->chromosomes(), f->resolution(), c.path_to_bin_mask);
 
   using File = std::remove_cvref_t<decltype(*f)>;
-  const auto evs = ExpectedValues<File>::trans_only(
+  const auto evs = ExpectedValues::trans_only(
       f,
       {c.mad_max, c.min_delta, c.max_delta, c.bin_aggregation_possible_distances_cutoff,
        c.bin_aggregation_observed_distances_cutoff, c.interpolate_expected_values,
@@ -109,7 +109,7 @@ static void process_one_chromosome_pair(FilePtr f, const ExpectedConfig &c) {
   const auto &chrom2 = f->chromosomes().at(c.chrom2);
 
   using File = std::remove_cvref_t<decltype(*std::declval<FilePtr>())>;
-  const auto evs = ExpectedValues<File>::chromosome_pair(
+  const auto evs = ExpectedValues::chromosome_pair(
       f, chrom1, chrom2,
       {c.mad_max, c.min_delta, c.max_delta, c.bin_aggregation_possible_distances_cutoff,
        c.bin_aggregation_observed_distances_cutoff, c.interpolate_expected_values,
@@ -121,45 +121,23 @@ static void process_one_chromosome_pair(FilePtr f, const ExpectedConfig &c) {
   evs.serialize(c.output_path);
 }
 
-// clang-format off
-  using HiCFilePtr =
-      std::variant<
-          std::shared_ptr<const hictk::cooler::File>,
-          std::shared_ptr<const hictk::hic::File>>;
-// clang-format on
-
-[[nodiscard]] static HiCFilePtr open_file_ptr(const std::filesystem::path &path,
-                                              std::uint32_t resolution) {
-  hictk::File f_(path.string(), resolution);
-  return {std::visit(
-      [&](auto &&ff) {
-        using FileT = std::remove_reference_t<decltype(ff)>;
-        return HiCFilePtr{std::make_shared<const FileT>(std::forward<FileT>(ff))};
-      },
-      f_.get())};
-}
-
 int run_nchg_expected(const ExpectedConfig &c) {
-  const auto f = open_file_ptr(c.input_path, c.resolution);
+  const auto f = std::make_shared<const hictk::File>(c.input_path.string(), c.resolution);
 
-  std::visit(
-      [&](const auto &f_) {
-        if (c.cis_only) {
-          process_cis_chromosomes(f_, c);
-          return;
-        }
-        if (c.trans_only) {
-          process_trans_chromosomes(f_, c);
-          return;
-        }
-        if (c.chrom1 == "all") {
-          assert(c.chrom2 == "all");
-          process_all_chromosomes(f_, c);
-          return;
-        }
-        process_one_chromosome_pair(f_, c);
-      },
-      f);
+  if (c.cis_only) {
+    process_cis_chromosomes(f, c);
+    return 0;
+  }
+  if (c.trans_only) {
+    process_trans_chromosomes(f, c);
+    return 0;
+  }
+  if (c.chrom1 == "all") {
+    assert(c.chrom2 == "all");
+    process_all_chromosomes(f, c);
+    return 0;
+  }
+  process_one_chromosome_pair(f, c);
 
   return 0;
 }
